Move `flang_rt_verbose_abort` to new file `stl-overrides.cpp`. Include that new file in the build rules. Change introductory paragraph in `stl-overrides.h`.
