+++ flang-rt/lib/runtime/io-api-minimal.cpp
@@ -146,18 +146,4 @@ bool IODEF(OutputLogical)(Cookie cookie, bool truth) {
 
 } // namespace Fortran::runtime::io
 
-#if defined(_LIBCPP_VERBOSE_ABORT)
-// Provide function that is used in place of `std::__libcpp_verbose_abort` to
-// avoid dependency on the symbol provided by libc++.
-
-void flang_rt_verbose_abort(char const *format, ...) {
-  va_list list;
-  va_start(list, format);
-  std::vfprintf(stderr, format, list);
-  va_end(list);
-
-  std::abort();
-}
-#endif
-
 RT_EXT_API_GROUP_END
